Move syscall types needed by shim into shadow-shim-helper-rs
C code in the shadow-shim-helper-rs crate that depends on SysCallReg etc
was failing to compile in clean builds, since those definitions are now
in the dynamically generated bindings header instead of statically
present.

Moving these definitions into shadow-shim-helper-rs fixes this
dependency issue, and will ultimately be needed anyway to implement more
shim and shim-shared-memory code in Rust.

Compilation currently fails due to conflicting trait implementations;
fixed in next commit.
