+++ src/lib/shadow-shim-helper-rs/src/lib.rs
@@ -12,6 +12,7 @@ pub mod rootedcell;
 pub mod shim_shmem;
 pub mod signals;
 pub mod simulation_time;
+pub mod syscall_types;
 
 #[repr(transparent)]
 #[derive(
