+++ src/main/build.rs
@@ -100,8 +100,9 @@ fn run_cbindgen(build_common: &ShadowBuildCommon) {
         .with_crate(crate_dir)
         .with_config(cbindgen::Config {
             // We want to avoid including any C headers from this crate here,
-            // which lets us avoid circular dependencies.
-            includes: vec![],
+            // which lets us avoid circular dependencies. Ok to depend on headers
+            // generated by crates that this one depends on.
+            includes: vec!["lib/shadow-shim-helper-rs/shim_helper.h".into()],
             include_guard: Some("main_opaque_bindings_h".into()),
             after_includes: {
                 let mut v = base_config.after_includes.clone().unwrap();
@@ -265,6 +266,10 @@ fn run_bindgen(build_common: &ShadowBuildCommon) {
         .blocklist_type("QDiscMode")
         // Imported from libc crate below
         .blocklist_type("siginfo_t")
+        .blocklist_type("SysCallReg")
+        .blocklist_type("SysCallArgs")
+        .blocklist_type("PluginPtr")
+        .blocklist_type("PluginPhysicalPtr")
         .disable_header_comment()
         .raw_line("/* automatically generated by rust-bindgen */")
         .raw_line("use libc::siginfo_t;")
@@ -278,13 +283,14 @@ fn run_bindgen(build_common: &ShadowBuildCommon) {
         .raw_line("use crate::host::host::Host;")
         .raw_line("use crate::host::process::ProcessRefCell;")
         .raw_line("use crate::host::syscall::handler::SyscallHandler;")
-        .raw_line("use crate::host::syscall_types::{PluginPtr, PluginPhysicalPtr, SysCallArgs, SysCallReturnBody};")
+        .raw_line("use crate::host::syscall_types::SysCallReturnBody;")
         .raw_line("use crate::host::thread::Thread;")
         .raw_line("use crate::utility::counter::Counter;")
         .raw_line("use crate::utility::legacy_callback_queue::RootedRefCell_StateEventSource;")
         .raw_line("")
         .raw_line("use logger::Logger;")
         .raw_line("use shadow_shim_helper_rs::HostId;")
+        .raw_line("use shadow_shim_helper_rs::syscall_types::{PluginPtr, PluginPhysicalPtr, SysCallArgs};")
         .raw_line("#[allow(unused)]")
         .raw_line("use shadow_shim_helper_rs::shim_shmem::{HostShmem, HostShmemProtected, ProcessShmem, ThreadShmem};")
         .raw_line("#[allow(unused)]")
